+++ src/app/shared/components/ride-info/ride-info.component.scss
@@ -1,4 +1,4 @@
-@import '@taiga-ui/core/styles/taiga-ui-local';
+@import '../../../../../node_modules/@taiga-ui/core/styles/taiga-ui-local';
 
 $height-line: 5rem;
 $shift-up: -0.8rem;
