+++ src/app/shared/utils/calculate-train-stop-duration.ts
